refactor: reorganize question components into components/ subfolders with descriptive names
