+++ packages/web/src/features/questions/multi/index.ts
@@ -1,11 +1,11 @@
 import { SCORING_MODES } from "@razzia/common/constants"
 import type { MultiQuestionOptions } from "@razzia/common/types/game"
 
-export { default as AnswerComponent } from "@razzia/web/features/questions/multi/AnswerComponent"
+export { default as AnswerComponent } from "@razzia/web/features/questions/multi/components/MultiAnswers"
 
-export { default as ConfigComponent } from "@razzia/web/features/questions/multi/ConfigComponent"
+export { default as ConfigComponent } from "@razzia/web/features/questions/multi/components/MultiConfig"
 
-export { default as SolutionPicker } from "@razzia/web/features/questions/multi/SolutionPicker"
+export { default as SolutionPicker } from "@razzia/web/features/questions/multi/components/MultiPicker"
 
 export const labelKey = "quizz:questionType.multi"
